+++ clowder/clowder_repo.py
@@ -3,23 +3,10 @@ import atexit
 import os
 import sys
 from termcolor import colored
-from clowder.utility.git_utilities import (
-    git_add,
-    git_branches,
-    git_checkout,
-    git_commit,
-    git_create_repo,
-    git_existing_repository,
-    git_fetch_silent,
-    git_is_dirty,
-    git_pull,
-    git_push,
-    git_reset_head,
-    git_status,
-    git_validate_repo
-)
+from clowder.utility.git_utilities import Git
 from clowder.utility.clowder_utilities import (
     execute_command,
+    existing_git_repository,
     force_symlink,
     is_offline,
     remove_directory_exit
@@ -47,39 +34,45 @@ class ClowderRepo(object):
 
     def add(self, files):
         """Add files in clowder repo to git index"""
-        git_add(self.clowder_path, files)
-        git_status(self.clowder_path)
+        repo = Git(self.clowder_path)
+        repo.add(files)
+        repo.status()
 
     def branches(self):
         """Return current local branches"""
-        return git_branches(self.clowder_path)
+        repo = Git(self.clowder_path)
+        return repo.branches()
 
     def checkout(self, ref):
         """Checkout ref in clowder repo"""
+        repo = Git(self.clowder_path)
         if self.is_dirty():
             print(' - Dirty repo. Please stash, commit, or discard your changes')
-            git_status(self.clowder_path)
+            repo.status()
         else:
-            git_checkout(self.clowder_path, ref)
+            repo.checkout(ref)
 
     def clean(self):
         """Discard changes in clowder repo"""
+        repo = Git(self.clowder_path)
         if self.is_dirty():
             print(' - Discard current changes')
-            git_reset_head(self.clowder_path)
+            repo.reset_head()
         else:
             print(' - No changes to discard')
 
     def commit(self, message):
         """Commit current changes in clowder repo"""
-        git_commit(self.clowder_path, message)
+        repo = Git(self.clowder_path)
+        repo.commit(message)
 
     def init(self, url, branch):
         """Clone clowder repo from url"""
         # Register exit handler to remove files if cloning repo fails
         atexit.register(self.init_exit_handler)
         repo_branch = 'refs/heads/' + branch
-        git_create_repo(self.clowder_path, url, 'origin', repo_branch)
+        repo = Git(self.clowder_path)
+        repo.create_repo(url, 'origin', repo_branch)
         self.link()
 
     def init_exit_handler(self):
@@ -91,7 +84,8 @@ class ClowderRepo(object):
 
     def is_dirty(self):
         """Check if project is dirty"""
-        return git_is_dirty(self.clowder_path)
+        repo = Git(self.clowder_path)
+        return repo.is_dirty(self.clowder_path)
 
     def link(self, version=None):
         """Create symlink pointing to clowder.yaml file"""
@@ -113,13 +107,14 @@ class ClowderRepo(object):
     def print_status(self, fetch=False):
         """Print clowder repo status"""
         repo_path = os.path.join(self.root_directory, '.clowder')
-        if not git_existing_repository(repo_path):
+        repo = Git(self.clowder_path)
+        if not existing_git_repository(repo_path):
             output = colored('.clowder', 'green')
             print(output)
             return
         if not is_offline() and fetch:
             print(' - Fetch upstream changes for clowder repo')
-            git_fetch_silent(self.clowder_path)
+            repo.fetch_silent()
         project_output = format_project_string(repo_path, '.clowder')
         current_ref_output = format_project_ref_string(repo_path)
 
@@ -137,11 +132,13 @@ class ClowderRepo(object):
 
     def pull(self):
         """Pull clowder repo upstream changes"""
-        git_pull(self.clowder_path)
+        repo = Git(self.clowder_path)
+        repo.pull()
 
     def push(self):
         """Push clowder repo changes"""
-        git_push(self.clowder_path)
+        repo = Git(self.clowder_path)
+        repo.push()
 
     def run_command(self, command):
         """Run command in clowder repo"""
@@ -153,11 +150,13 @@ class ClowderRepo(object):
 
     def status(self):
         """Print clowder repo git status"""
-        git_status(self.clowder_path)
+        repo = Git(self.clowder_path)
+        repo.status()
 
     def _validate_groups(self):
         """Validate status of clowder repo"""
-        if not git_validate_repo(self.clowder_path):
+        repo = Git(self.clowder_path)
+        if not repo.validate_repo():
             print_validation(self.clowder_path)
             print()
             sys.exit(1)