+++ clowder/project.py
@@ -5,6 +5,7 @@ from termcolor import cprint
 from clowder.fork import Fork
 from clowder.utility.clowder_utilities import (
     execute_forall_command,
+    existing_git_repository,
     is_offline
 )
 from clowder.utility.print_utilities import (
@@ -20,36 +21,7 @@ from clowder.utility.git_print_utilities import (
     print_exists,
     print_validation
 )
-from clowder.utility.git_utilities import (
-    git_abort_rebase,
-    git_clean,
-    git_clean_all,
-    git_configure_remotes,
-    git_existing_local_branch,
-    git_existing_remote_branch,
-    git_existing_repository,
-    git_fetch_remote,
-    git_has_submodules,
-    git_herd,
-    git_herd_branch,
-    git_herd_branch_upstream,
-    git_herd_upstream,
-    git_print_branches,
-    git_prune_local,
-    git_prune_remote,
-    git_is_rebase_in_progress,
-    git_reset_head,
-    git_sha_long,
-    git_start,
-    git_start_offline,
-    git_stash,
-    git_status,
-    git_submodules_clean,
-    git_submodules_reset,
-    git_submodules_update,
-    git_sync,
-    git_validate_repo
-)
+from clowder.utility.git_utilities import Git
 
 # Disable errors shown by pylint for too many public methods
 # pylint: disable=R0904
@@ -129,14 +101,15 @@ class Project(object):
         if not os.path.isdir(self.full_path()):
             cprint(" - Project is missing\n", 'red')
             return
+        repo = Git(self.full_path())
         if not is_offline():
             if remote:
                 if self.fork is None:
-                    git_fetch_remote(self.full_path(), self.remote_name, self.depth)
+                    repo.fetch(self.remote_name, depth=self.depth)
                 else:
-                    git_fetch_remote(self.full_path(), self.fork.remote_name, 0)
-                    git_fetch_remote(self.full_path(), self.remote_name, 0)
-        git_print_branches(self.full_path(), local=local, remote=remote)
+                    repo.fetch(self.fork.remote_name)
+                    repo.fetch(self.remote_name)
+        repo.print_branches(local=local, remote=remote)
 
     def clean(self, args=None, recursive=False):
         """Discard changes for project"""
@@ -144,20 +117,21 @@ class Project(object):
         if not os.path.isdir(self.full_path()):
             cprint(" - Project is missing\n", 'red')
             return
+        repo = Git(self.full_path())
         print(' - Clean project')
-        git_clean(self.full_path(), args=args)
+        repo.clean(args=args)
         print(' - Reset project')
-        git_reset_head(self.full_path())
-        if git_is_rebase_in_progress(self.full_path()):
+        repo.reset_head()
+        if repo.is_rebase_in_progress():
             print(' - Abort rebase in progress')
-            git_abort_rebase(self.full_path())
+            repo.abort_rebase()
         if self.recursive and recursive:
             print(' - Clean submodules recursively')
-            git_submodules_clean(self.full_path())
+            repo.submodules_clean()
             print(' - Reset submodules recursively')
-            git_submodules_reset(self.full_path())
+            repo.submodules_reset()
             print(' - Update submodules recursively')
-            git_submodules_update(self.full_path())
+            repo.submodules_update()
 
     def clean_all(self):
         """Discard all changes for project"""
@@ -165,20 +139,21 @@ class Project(object):
         if not os.path.isdir(self.full_path()):
             cprint(" - Project is missing\n", 'red')
             return
+        repo = Git(self.full_path())
         print(' - Clean project')
-        git_clean_all(self.full_path())
+        repo.clean(args='fdx')
         print(' - Reset project')
-        git_reset_head(self.full_path())
-        if git_is_rebase_in_progress(self.full_path()):
+        repo.reset_head()
+        if repo.is_rebase_in_progress():
             print(' - Abort rebase in progress')
-            git_abort_rebase(self.full_path())
+            repo.abort_rebase()
         if self.recursive:
             print(' - Clean submodules recursively')
-            git_submodules_clean(self.full_path())
+            repo.submodules_clean()
             print(' - Reset submodules recursively')
-            git_submodules_reset(self.full_path())
+            repo.submodules_reset()
             print(' - Update submodules recursively')
-            git_submodules_update(self.full_path())
+            repo.submodules_update()
 
     def diff(self):
         """Show git diff for project"""
@@ -186,7 +161,8 @@ class Project(object):
         if not os.path.isdir(self.full_path()):
             cprint(" - Project is missing\n", 'red')
             return
-        git_status(self.full_path())
+        repo = Git(self.full_path())
+        repo.status()
 
     def exists(self):
         """Check if project exists on disk"""
@@ -195,23 +171,25 @@ class Project(object):
 
     def existing_branch(self, branch, is_remote):
         """Check if branch exists"""
+        repo = Git(self.full_path())
         if is_remote:
             if self.fork is None:
-                return git_existing_remote_branch(self.full_path(), branch, self.remote_name)
+                return repo.existing_remote_branch(branch, self.remote_name)
             else:
-                return git_existing_remote_branch(self.full_path(), branch, self.fork.remote_name)
+                return repo.existing_remote_branch(branch, self.fork.remote_name)
         else:
-            return git_existing_local_branch(self.full_path(), branch)
+            return repo.existing_local_branch(branch)
 
     def fetch_all(self):
         """Fetch upstream changes if project exists on disk"""
         self._print_status()
+        repo = Git(self.full_path())
         if self.exists():
             if self.fork is None:
-                git_fetch_remote(self.full_path(), self.remote_name, self.depth)
+                repo.fetch(self.remote_name, depth=self.depth)
             else:
-                git_fetch_remote(self.full_path(), self.fork.remote_name, 0)
-                git_fetch_remote(self.full_path(), self.remote_name, 0)
+                repo.fetch(self.fork.remote_name)
+                repo.fetch(self.remote_name)
         else:
             self.print_exists()
 
@@ -229,7 +207,8 @@ class Project(object):
         if resolved:
             ref = self.ref
         else:
-            ref = git_sha_long(self.full_path())
+            repo = Git(self.full_path())
+            ref = repo.sha_long()
         project = {'name': self.name,
                    'path': self.path,
                    'depth': self.depth,
@@ -249,48 +228,51 @@ class Project(object):
         else:
             herd_depth = depth
 
+        repo = Git(self.full_path())
         if branch is None:
             if self.fork is None:
                 self._print_status()
-                git_herd(self.full_path(), self.url, self.remote_name, self.ref,
-                         depth=herd_depth, recursive=self.recursive)
+                repo.herd(self.url, self.remote_name, self.ref,
+                              depth=herd_depth, recursive=self.recursive)
             else:
                 self.fork.print_status()
-                git_configure_remotes(self.full_path(), self.remote_name, self.url,
-                                      self.fork.remote_name, self.fork.url)
+                repo.configure_remotes(self.remote_name, self.url,
+                                           self.fork.remote_name, self.fork.url)
                 print(format_fork_string(self.fork.name))
-                git_herd(self.full_path(), self.fork.url, self.fork.remote_name,
-                         self.ref, recursive=self.recursive)
+                repo.herd(self.fork.url, self.fork.remote_name,
+                              self.ref, recursive=self.recursive)
                 print(format_fork_string(self.name))
-                git_herd_upstream(self.full_path(), self.url, self.remote_name, self.ref)
+                repo.herd_upstream(self.url, self.remote_name, self.ref)
         else:
             if self.fork is None:
                 self._print_status()
-                git_herd_branch(self.full_path(), self.url, self.remote_name,
-                                branch, self.ref, depth=herd_depth, recursive=self.recursive)
+                repo.herd_branch(self.url, self.remote_name,
+                                     branch, self.ref, depth=herd_depth, recursive=self.recursive)
             else:
                 self.fork.print_status()
-                git_configure_remotes(self.full_path(), self.remote_name, self.url,
-                                      self.fork.remote_name, self.fork.url)
+                repo.configure_remotes(self.remote_name, self.url,
+                                           self.fork.remote_name, self.fork.url)
                 print(format_fork_string(self.fork.name))
-                git_herd_branch(self.full_path(), self.fork.url, self.fork.remote_name,
-                                branch, self.ref, recursive=self.recursive)
+                repo.herd_branch(self.fork.url, self.fork.remote_name,
+                                     branch, self.ref, recursive=self.recursive)
                 print(format_fork_string(self.name))
-                git_herd_branch_upstream(self.full_path(), self.url, self.remote_name,
-                                         branch, self.ref)
+                repo.herd_branch_upstream(self.url, self.remote_name,
+                                              branch, self.ref)
 
     def is_dirty(self):
         """Check if project is dirty"""
-        if not git_validate_repo(self.full_path()):
+        repo = Git(self.full_path())
+        if not repo.validate_repo():
             return True
         if self.recursive:
-            if git_has_submodules(self.full_path()):
+            if repo.has_submodules():
                 return True
         return False
 
     def is_valid(self):
         """Validate status of project"""
-        return git_validate_repo(self.full_path())
+        repo = Git(self.full_path())
+        return repo.validate_repo()
 
     def print_exists(self):
         """Print existence validation message for project"""
@@ -306,7 +288,7 @@ class Project(object):
 
     def prune(self, branch, force=False, local=False, remote=False):
         """Prune branch"""
-        if not git_existing_repository(self.full_path()):
+        if not existing_git_repository(self.full_path()):
             return
         if local and remote:
             self._prune_local(branch, force)
@@ -342,7 +324,8 @@ class Project(object):
     def start(self, branch, tracking):
         """Start a new feature branch"""
         self._print_status()
-        if not git_existing_repository(self.full_path()):
+        repo = Git(self.full_path())
+        if not existing_git_repository(self.full_path()):
             cprint(" - Directory doesn't exist", 'red')
             return
         if self.fork is None:
@@ -352,9 +335,9 @@ class Project(object):
             remote = self.fork.remote_name
             depth = 0
         if is_offline():
-            git_start_offline(self.full_path(), branch)
+            repo.start_offline(branch)
         else:
-            git_start(self.full_path(), remote, branch, depth, tracking)
+            repo.start(remote, branch, depth, tracking)
 
     def status(self, padding):
         """Print status for project"""
@@ -364,37 +347,37 @@ class Project(object):
         """Stash changes for project if dirty"""
         if self.is_dirty():
             self._print_status()
-            git_stash(self.full_path())
+            repo = Git(self.full_path())
+            repo.stash()
 
     def sync(self):
         """Print status for project"""
         self.fork.print_status()
-        git_configure_remotes(self.full_path(), self.remote_name, self.url,
-                              self.fork.remote_name, self.fork.url)
+        repo = Git(self.full_path())
+        repo.configure_remotes(self.remote_name, self.url,
+                                   self.fork.remote_name, self.fork.url)
         print(format_fork_string(self.fork.name))
-        git_herd(self.full_path(), self.fork.url, self.fork.remote_name,
-                 self.ref, recursive=self.recursive)
+        repo.herd(self.fork.url, self.fork.remote_name,
+                      self.ref, recursive=self.recursive)
         print(format_fork_string(self.name))
-        git_herd_upstream(self.full_path(), self.url, self.remote_name,
-                          self.ref)
+        repo.herd_upstream(self.url, self.remote_name, self.ref)
         self.fork.print_status()
-        git_sync(self.full_path(), self.remote_name, self.fork.remote_name,
-                 self.ref, self.recursive)
+        repo.sync(self.remote_name, self.fork.remote_name,
+                      self.ref, self.recursive)
 
     def _print_status(self):
         """Print formatted project status"""
-        repo_path = os.path.join(self.root_directory, self.path)
-        if not git_existing_repository(repo_path):
+        if not existing_git_repository(self.full_path()):
             cprint(self.path, 'green')
             return
-        project_output = format_project_string(repo_path, self.path)
-        current_ref_output = format_project_ref_string(repo_path)
+        project_output = format_project_string(self.full_path(), self.path)
+        current_ref_output = format_project_ref_string(self.full_path())
         print(project_output + ' ' + current_ref_output)
 
     def _print_status_indented(self, padding):
         """Print formatted and indented project status"""
         repo_path = os.path.join(self.root_directory, self.path)
-        if not git_existing_repository(repo_path):
+        if not existing_git_repository(self.full_path()):
             cprint(self.name, 'green')
             return
         project_output = format_project_string(repo_path, self.path)
@@ -403,9 +386,10 @@ class Project(object):
 
     def _prune_local(self, branch, force):
         """Prune local branch"""
-        if git_existing_local_branch(self.full_path(), branch):
+        repo = Git(self.full_path())
+        if repo.existing_local_branch(branch):
             self._print_status()
-            git_prune_local(self.full_path(), branch, self.ref, force)
+            repo.prune_local(branch, self.ref, force)
 
     def _prune_remote(self, branch):
         """Prune remote branch"""
@@ -413,6 +397,7 @@ class Project(object):
             remote = self.remote_name
         else:
             remote = self.fork.remote_name
-        if git_existing_remote_branch(self.full_path(), branch, remote):
+        repo = Git(self.full_path())
+        if repo.existing_remote_branch(branch, remote):
             self._print_status()
-            git_prune_remote(self.full_path(), branch, remote)
+            repo.prune_remote(branch, remote)